added default admin user and change password
many minor code cleanups and fixes

+++ UserIdentity.Api/Handlers/UserHandlers/DeleteUserHandler.cs
@@ -44,6 +44,12 @@ namespace UserIdentity.Api.Handlers.UserHandlers
                 throw new RpcException(new Status(StatusCode.NotFound, "User not found"));
             }
 
+            if (user.UserName == "admin")
+            {
+                throw new RpcException(new Status(StatusCode.PermissionDenied,
+                    "Default admin account cannot be deleted!"));
+            }
+
             var result = await _userManager.DeleteAsync(user);
             if (!result.Succeeded)
             {
